DERBY-1793
merging change 439041 from trunk into 10.2 branch:

Increasing the maximum time to wait for the server to start up from 30 to 60
seconds.  The test checks in 500ms increments, so this change does not increase
the time for test to run for those who are already successfully running it.
So far this change has made this test pass in my environment, where previously
it failed consistently.  Since this test passes in the nightly full test runs
across a number of environments I assume the network server startup time is
somehow related to my particular machine (processor, memory, disk frag, firewal
, vpn, ...)

Others have seen this issue so I am committing to the codeline.  I have filed
a separate issue that work should be done to measure the performance of
network server startup as a targeted test (DERBY-1794).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@439051 13f79535-47bb-0310-9956-ffa450edef68