DERBY-1878 contributed by Sunitha Kambhampati merging test only change 450508 from trunk to 10.2 branch
I am attaching the patch (derby1878.diff.txt) to improve some error handling in
some of the network server tests.

1.The execCmdDumpResults used by the five tests timeslice.java,maxthreads.java,
testij.java,runtimeinfo.java,sysinfo.java suffer from the same problems that was
fixed for testProperties.java namely
-- the outputstream for the sub process's is not flushed out
-- there is no timeout handling for the ProcessStreamResult
2.Eliminate duplication of code in these 5 tests for execCmdDumpResults(String[
] args) method. The execCmdDumpResults method basically exec's a new process and
then waits for the process to dump the results. A new utility class - ExecProcU
til is added with execCmdDumpResults that is generalized to take the necessary i
nput from the tests as well as fixes the issues mentioned in #1. The OutputStrea
m is flushed out by calling bos.flush and System.out.flush and the timeout is ad
ded for ProcessStreamResult.
3.Make use of the TimedProcess to kill process if process does not exit within
the timeout period.
4.TestConnection.java test has some variation of the execCmdDumpResults and it
also adds some testcases into this method. Hence this method in this test is lef
t as is and timeout handling is added.
5.testij.out has been updated. The previous master file was eating away the las
t line that was written to System.out, but now that the process's streams are fl
ushed properly, the last line in testij.out test which prints out 'End test'
is also seen in the output file.

Also noticed that these tests - like timeslice.java, maxthreads.java all seem t
o set the properties for the server and then check if the property is set. The f
unctionality of the server when these properties is set is not being tested. It
will be good to add tests that test that the functionality itself is working as
expected or not. Please note, recently connecting to server with timeslice optio
ns discovered some issues with server (see DERBY-1856).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@450884 13f79535-47bb-0310-9956-ffa450edef68