DERBY-1925 : Add re-encrytion of database test cases to the upgrade test.
This patch adds test cases to the upgrade test to test encryption of an 
un-encrypted database and re-encryption of encrypted database with new 
password on soft-upgrade and hard-upgrade from 10.1 to 10.2. These new test 
cases will be executed only when the test is run using jars, becuase currently
upgrade test setup mixes two version of derby engine classes when run using 
the classes in a directory; this cases boot time error when database is
encrypted. Derby engine is not expected to work when two version of derby 
classes are in the class path.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@452682 13f79535-47bb-0310-9956-ffa450edef68