DERBY-1847 contributed by Mamta Satoor backporting via merge change 453352 from trunk to 10.2 branch.
To recap the problem, in SQL Authorization mode, when a new column is added to
a table, the rows in SYSCOLPERMS for the table in question were not getting 
updated to incorporate the new column. This caused ASSERT failure when a 
non-table owner attempted to select the new column.

Some background information on system table involved: SYSCOLPERMS keeps track 
of column level privileges on a given table. One of the columns in SYSCOLPERMS
is "COLUMNS" and it has a bit map to show which columns have the given 
permission granted on them. When a new column is added to the user table, 
the "COLUMNS" need to be expanded by one bit and that bit should be initialized
to zero since no privileges have been granted on that column at the ALTER 
TABLE...ADD COLUMN time.

I have fixed this problem by having 
AlterTableConstantAction.addNewColumnToTable call the new method in 
DataDictionary called updateSYSCOLPERMSforAddColumnToUserTable. At this point, 
we know of only the TableDescriptor's uuid which can help us determine all the 
rows in SYSCOLPERMS for that given table uuid. I get ColPermsDescriptor for 
each one of those rows and then use the ColPermsDescriptor's uuid to update 
the "COLUMNS" column so SYSCOLPERMS is aware of the newly added column in user
table. This fixes the problem because at the time of SELECT, when we do 
privilege lookup in SYSCOLPERMS, we have info on the newly added column.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@453747 13f79535-47bb-0310-9956-ffa450edef68