DERBY-2371: Altering VARCHAR default fails when column contains data
Merged trunk revision 512019 to the 10.2 branch. Description of change:

Using ALTER TABLE to modify the default value for a VARCHAR column when
the column already contains data throws ERROR 22018: Invalid character
string format for type long. 

The problem arises in AlterTableConstantAction.modifyColumnDefault, which
only understands two types of modifications that can be made to a
column's default value:
 - changing the INCREMENT BY value for a generated IDENTITY column
 - changing the RESTART WITH value for a generated IDENTITY column

But there is a third type of modification that can be made to a column's
default value, which is simply to change the default value that is to
be used for an ordinary column if no value is provided when inserting
a new row into that table.

The fix is to add a new ColumnInfo modification type code:
   MODIFY_COLUMN_DEFAULT_VALUE
and to teach ModifyColumnNode and AlterTableConstantAction what to do
for modifications of type MODIFY_COLUMN_DEFAULT_VALUE. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@512334 13f79535-47bb-0310-9956-ffa450edef68