DERBY-2376: Make classpath entries portable in Eclipse plug-in
This patch was contributed by Aaron Tarter (ajtarter@gmail.com)

This patch merges the DERBY-2376 changes from the trunk to the 10.2 branch.
The two revisions merged were: 515864 and 531595. svn merge -r was used.

The 515864 revision contains a change to DerbyUtils to manage the classpath
entries relative to ECLIPSE_HOME. The 531595 revision contains modifications
to the Eclipse plug-in documentation to provide modified screen-shots.
The new screen-shot images show the new format of the classpath entries,
relative to ECLIPSE_HOME.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@531597 13f79535-47bb-0310-9956-ffa450edef68