DERBY-2549, backporting change 540657 from trunk to 10.2 line. contributed by Mayuresh Nirhali
Fix null pointer when running inplace compress.  Change code to correctly
handle when more than 100 rows are moved from a single page.  The new code
returns to the caller after processing the 100 rows, and the next trip
through the loop picks up the scan where it left off on that same page.
Test case was added to existing test.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@540743 13f79535-47bb-0310-9956-ffa450edef68