DERBY-4331 : Join returns results in wrong order
backporting fix (#801481) from trunk to 10.2 branch.

Fixes a number of sort avoidance bugs that were introduced by the fix
for DERBY-3926.  This check in backs out the equi-join part of the DERBY-3926.
The changes for this were isolated and were the only changes to
FromBaseTable.java.  Backing out only this part of the 3926 checkin fixes new
problems identified in DERBY-4331, and continues to fix the problem queries in
DERBY-3926.  Knowledge of equijoin is no longer used as a factor for
sort avoidance.

Also included is an update to the wisconsin tests.  A number of diffs
resulted from different join order to maintain a sort avoidance plan.  2
queries identified in DERBY-4339 no longer use sort avoidance.

Test cases in the junit OrderByAndSortAvoidance test are not backported due
to limited junit support in the 10.2 branch.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@802958 13f79535-47bb-0310-9956-ffa450edef68