DERBY-3337 convert jdbcapi/derbyStress.java to JUnit DERBY-4200 client side OutOfMemoryError running derbnetclientmats:jdbcapi/derbyStress
Merged revision 1141924 from trunk. 

Convert test to JUnit and also run the finalizer periodically if freeMemory() falls below 500K.  The converted tests is no longer called derbyStress. It is now memory.MemoryLeakFixes and runs with the junit-lowmem target or can be run directly with junit.textui.TestRunner with -Xmx16M.  




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.6@1142539 13f79535-47bb-0310-9956-ffa450edef68