DERBY-1913 storetests/st_reclaim_longcol.java fails intermittently
backported fix #1243622 from 10.8 branch to 10.6 branch.

The test was counting on being able to control the number of FREE pages,
but the number is very dependent on ability of background thread to run
in a timely manner. Changed the test to check number of allocated pages,
which at least correctly tests that what we think should be background reclaimed
eventually is. Still needs some wait logic which I think will work better
now. I tested against knut's patch to delay the daemon 1 second for every
piece of work and the old test always failed and the new test succeeded. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.6@1245681 13f79535-47bb-0310-9956-ffa450edef68