DERBY-5494 Same value returned by successive calls to a sequence generator flanking an unorderly shutdown. DERBY-5780 identity column performance has degredated
backporting change #1344065 from trunk to 10.9 branch.

The previous patch for DERBY-5494 had the unintended affect of forcing a
synchronous write for all nested user transactions at abort time.  This
in turn caused identity column inserts to have one synchronous write per
insert as the nested user transaction is destroyed for each insert which
does an abort each time.

To solve this interfaces were changed so that calling code could set the
default commit sync behavior when the transaction was committed rather than
count on the "type" of transaction.  Nested user transactions used for identity
columns have default set to not sync, and the rest of the nested user 
transactions default to syncing.  Behavior of other types of transactions 
should not be affected.  User transactions still sync by default and internal 
and ntt's still default to not sync.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.9@1344858 13f79535-47bb-0310-9956-ffa450edef68