DERBY-6131 select from view with "upper" and "in" list throws a ClassCastException
backporting change #1464594 from trunk to 10.9 branch.

Prior to this fix the following query on a view would throw a ClassCastException    select name from myView where upper(name) in ('AA', 'BB');
This query got in code that was trying to "push" predicates down, and
that code is only currently build to push constants and column references.
In this case it was mistakenly trying to push down the "upper" call, and
when it tried to cast it as a column reference it got the error.  The fix
was to check the type of the node and not push in this case.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.9@1465265 13f79535-47bb-0310-9956-ffa450edef68