DERBY-6237(PreparedStatement.execute() fails starting 10.2 when multiple rows are updated and PreparedStatement.setCharacterStream(int, Reader, int) is used)
Adding test for PreparedStatement.setCharacterStream on CLOB column and using that prepared statement to update more than one row of the table

In 10.1, setCharacterStream to update CLOB and varchar columns work even when update is going to update more than one row but in 10.2, PreparedStatement.setCharacterStream on CLOB column fails when running in embedded mode. Client server case works fine on 10.2. I have commented out the test for embedded in 10.2 Once this jira is fixed in 10.2, we should uncomment these tests.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@1488411 13f79535-47bb-0310-9956-ffa450edef68