DERBY-6237(PreparedStatement.execute() fails starting 10.2 when multiple rows are updated and PreparedStatement.setCharacterStream(int, Reader, int) is used)
This commit adds test case for updating more than one row with setCharacterStream to update VARCHAR column in the underlying table. The test is commented out at this point since it does not work in 10.2. Once this jira is fixed in 10.2, we should uncomment these tests. This test case works fine in 10.1




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.2@1488453 13f79535-47bb-0310-9956-ffa450edef68